split collective signing in more files
